esp_modbus: Fix Wno-format issues

+++ freemodbus/common/esp_modbus_master.c
@@ -36,7 +36,7 @@ esp_err_t mbc_master_destroy(void)
     MB_MASTER_CHECK((error == ESP_OK),
                     error,
                     "Master destroy failure, error=(0x%x).",
-                    error);
+                    (int)error);
     return error;
 }
 
@@ -53,7 +53,7 @@ esp_err_t mbc_master_get_cid_info(uint16_t cid, const mb_parameter_descriptor_t*
     MB_MASTER_CHECK((error == ESP_OK),
                     error,
                     "Master get cid info failure, error=(0x%x).",
-                    error);
+                    (int)error);
     return error;
 }
 
@@ -73,7 +73,7 @@ esp_err_t mbc_master_get_parameter(uint16_t cid, char* name, uint8_t* value, uin
     MB_MASTER_CHECK((error == ESP_OK),
                     error,
                     "Master get parameter failure, error=(0x%x) (%s).",
-                    error, esp_err_to_name(error));
+                    (int)error, esp_err_to_name(error));
     return error;
 }
 
@@ -93,7 +93,7 @@ esp_err_t mbc_master_send_request(mb_param_request_t* request, void* data_ptr)
     MB_MASTER_CHECK((error == ESP_OK),
                     error,
                     "Master send request failure error=(0x%x) (%s).",
-                    error, esp_err_to_name(error));
+                    (int)error, esp_err_to_name(error));
     return ESP_OK;
 }
 
@@ -114,7 +114,7 @@ esp_err_t mbc_master_set_descriptor(const mb_parameter_descriptor_t* descriptor,
     MB_MASTER_CHECK((error == ESP_OK),
                     error,
                     "Master set descriptor failure, error=(0x%x) (%s).",
-                    error, esp_err_to_name(error));
+                    (int)error, esp_err_to_name(error));
     return ESP_OK;
 }
 
@@ -134,7 +134,7 @@ esp_err_t mbc_master_set_parameter(uint16_t cid, char* name, uint8_t* value, uin
     MB_MASTER_CHECK((error == ESP_OK),
                     error,
                     "Master set parameter failure, error=(0x%x) (%s).",
-                    error, esp_err_to_name(error));
+                    (int)error, esp_err_to_name(error));
     return ESP_OK;
 }
 
@@ -154,7 +154,7 @@ esp_err_t mbc_master_setup(void* comm_info)
     MB_MASTER_CHECK((error == ESP_OK),
                     error,
                     "Master setup failure, error=(0x%x) (%s).",
-                    error, esp_err_to_name(error));
+                    (int)error, esp_err_to_name(error));
     return ESP_OK;
 }
 
@@ -174,7 +174,7 @@ esp_err_t mbc_master_start(void)
     MB_MASTER_CHECK((error == ESP_OK),
                     error,
                     "Master start failure, error=(0x%x) (%s).",
-                    error, esp_err_to_name(error));
+                    (int)error, esp_err_to_name(error));
     return ESP_OK;
 }
